TOTP & TOS

+++ app/controllers/SessionsController.php
@@ -29,23 +29,27 @@ class SessionsController extends \lithium\action\Controller {
 						)
 				));
 				if($details['oneCode']===$this->request->data['loginpassword']){
-				
-					$totp = $this->request->data['totp'];
-					if($totp==""){
+					if($details["TOTP.Validate"]==1 && $details["TOTP.Login"]==true){
+						$totp = $this->request->data['totp'];
+						$ga = new GoogleAuthenticator();
+						if($totp==""){
+							Auth::clear('member');
+							Session::delete('default');
+						}else{
+							$checkResult = $ga->verifyCode($details['secret'], $totp, 2);		
+							if ($checkResult) {
+								Session::write('default',$default);
+								$user = Session::read('default');
+								return $this->redirect('ex::dashboard');
+							}else{
+								Auth::clear('member');
+								Session::delete('default');
+							}
+						}
+					}else{
 						Session::write('default',$default);
 						$user = Session::read('default');
 						return $this->redirect('ex::dashboard');
-					}else{
-						$ga = new GoogleAuthenticator();
-						$checkResult = $ga->verifyCode($details['secret'], $totp, 2);		
-						if ($checkResult) {
-							Session::write('default',$default);
-							$user = Session::read('default');
-							return $this->redirect('ex::dashboard');
-						}else{
-							Auth::clear('member');
-							Session::delete('default');
-						}
 					}
 				}else{
 					Auth::clear('member');
@@ -67,7 +71,7 @@ class SessionsController extends \lithium\action\Controller {
 	 public function delete() {
         Auth::clear('member');
 		Session::delete('default');
-        return $this->redirect('ex::dashboard');
+        return $this->redirect('/');
     }
 }
 ?>
